Add file for current SSL certificate error
Currently when trying to run a bundle install for Rails applications
there is a SSL certification error when some of the gems try to
install/update. This is preventing rake from being installed which
means that databases cannot be created/used. The solution for this
problem will be documented in the file for easier reference in the
future. It is also to document all problems that I encounter when
running Ruby/Rails on my Windows 10 system.