Define SSL Certification Error
This file was created because I am receiving a SSL Certification
error when trying to update any Ruby gem. However, the actual
error was not defined inside of the file. So I copied the exact
error that I am receiving from my command line for the problem to
be easily identified next time.