Add solution
A solution was implemented for my local environment which stopped
the error message from displaying and appropriately installs gems
from rubygems.org. This solution was not discussed in the file,
which is the whole point of the file, so all of the steps that were
followed to solve the problem were added.