Define postgres authentication problem
There is currently a problem when postgres database commands are run
on my Windows 10 machine. This is a new problem, because I used to
be able to run all of my Rails apps including their database
configurations. However, now there is an authentication problems
when the postgres commands are run. This file now defines the
entire outline of the problem that I have been running into.