Begin list of attempts to fix issue
I have tried and will continue to try multiple different ways to fix
the current authentication problem with postgresql. The first attempt
I made was added to the list which included trying to save my username
and password inside my database yaml file for the authentication, but
they did not work.