Add ENV method for pg username and password
Since the error message kept mentioning that there was no password
I added environmental variables to the database configuration that
called for the username and password that I setup using the PGAdmin.
However, after correctly implementing the variables I still receive
an error that says the username and password don't match, so it
still can't create any of the databases. Which isn't helpful.