Update README to include postgres issue
The previous README only had my the SSL Certification issue listed
on it. Since I came across a new issue and added a separate file for
its resolution, it needed to be added to the list that is displayed
on the main page of the repo. Therefore I added the postgres
authentication error to the page. The listing contains a brief
overview of the problems as well as the error message received when
performing the process. The message is displayed using a code block.