Add current status of issue - backlogged
This method was taking too long to solve my issue at hand so I had
to backlog the issue and use and quick, unsecure fix for it. While
all of the new files are still there and configured the same way,
I overrode the use of the ENV in the database yaml file by directly
using the login information instead. This fix is very unsecure and
I would not use it on someone else's project, it is okay for my
portfolio website and I will continue to work on the real solution
after I get some real project work done.